refactor: split completion failure reports
