+++ apps/backend/src/entrypoints/scheduledJobs/lambda-multipart-completion-reconciliation.ts
@@ -1,10 +1,12 @@
 import type { Handler } from "aws-lambda";
 import {
   MultipartCompletionFailureReportBatchError,
-  MultipartCompletionReconciliationBatchError,
   runMultipartCompletionFailureReportBatch,
-  runMultipartCompletionReconciliationBatch,
   type MultipartCompletionFailureReportBatchResult,
+} from "../../mediaAssets/multipart/completion/completionFailureReports";
+import {
+  MultipartCompletionReconciliationBatchError,
+  runMultipartCompletionReconciliationBatch,
   type MultipartCompletionReconciliationBatchResult,
 } from "../../mediaAssets/multipart/completion/completionReconciliation";
 import { createCloudWatchRecord } from "../../observability/cloudWatch";
