+++ apps/backend/src/mediaAssets/multipart/completion/completionReconciliation.ts
@@ -11,10 +11,7 @@ import {
   isTransientDatabaseError,
   TransientDatabaseHttpError,
 } from "../../../database/transient";
-import type {
-  BackendObservationScope,
-  MultipartCompletionReconciliationTerminalFailureDetails,
-} from "../../../observability/sentry";
+import type { BackendObservationScope } from "../../../observability/sentry";
 import { isLowercaseWorkspaceId } from "../../../workspaces/identity";
 import { mediaBlobCleanupDelayMs } from "../../blobLifecycle";
 import { isValidMediaAssetLastOperationId } from "../../lastOperationId";
@@ -35,6 +32,8 @@ import {
   type MediaAsset,
   type MediaBlobNormalizationVersion,
 } from "../../types";
+
+export * from "./completionFailureReports";
 
 const uuidPattern =
   /^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/u;
@@ -125,47 +124,6 @@ export type MultipartCompletionReconciliationDurableOutcome = Readonly<{
   errorCode: string | null;
 }>;
 
-export type ClaimedMultipartCompletionFailureReport = Readonly<{
-  failureEventId: string;
-  attemptToken: string;
-  workspaceId: string;
-  retryCount: number;
-  errorCode: string;
-  deliveryAttempt: number;
-  leaseToken: string;
-  leaseOwner: string;
-  leaseExpiresAt: string;
-}>;
-
-export type MultipartCompletionFailureReportBatchInput = Readonly<{
-  leaseOwner: string;
-  leaseDurationMs: number;
-  maximumReports: number;
-  deadlineAtMs: number;
-  reportTerminalFailure: (
-    details: MultipartCompletionReconciliationTerminalFailureDetails,
-  ) => void;
-  signal: AbortSignal;
-}>;
-
-export type MultipartCompletionFailureReportOutcome =
-  | "ambiguous"
-  | "lease_lost"
-  | "reported";
-
-export type MultipartCompletionFailureReportResult = Readonly<{
-  failureEventId: string;
-  outcome: MultipartCompletionFailureReportOutcome;
-}>;
-
-export type MultipartCompletionFailureReportBatchResult = Readonly<{
-  claimed: number;
-  ambiguous: number;
-  leaseLost: number;
-  reported: number;
-  results: ReadonlyArray<MultipartCompletionFailureReportResult>;
-}>;
-
 type ClaimedRow = Readonly<{
   attempt_token: string;
   reservation_token: string;
@@ -210,18 +168,6 @@ type DurableOutcomeRow = Readonly<{
   reconciliation_status: string;
   reconciliation_error_code: string | null;
 }>;
-type ClaimedFailureReportRow = Readonly<{
-  failure_event_id: string;
-  attempt_token: string;
-  workspace_id: string;
-  reconciliation_retry_count: number;
-  reconciliation_last_error_code: string;
-  failure_report_delivery_count: number;
-  failure_report_lease_token: string;
-  failure_report_lease_owner: string;
-  failure_report_lease_expires_at: Date;
-}>;
-
 export class MultipartCompletionReconciliationLeaseLostError extends Error {
   readonly code = "MULTIPART_COMPLETION_RECONCILIATION_LEASE_LOST";
 
@@ -230,17 +176,6 @@ export class MultipartCompletionReconciliationLeaseLostError extends Error {
       `Multipart completion reconciliation lease is no longer active. attemptToken=${attemptToken}`,
     );
     this.name = "MultipartCompletionReconciliationLeaseLostError";
-  }
-}
-
-export class MultipartCompletionFailureReportLeaseLostError extends Error {
-  readonly code = "MULTIPART_COMPLETION_FAILURE_REPORT_LEASE_LOST";
-
-  constructor(failureEventId: string) {
-    super(
-      `Multipart completion failure-report lease is no longer active. failureEventId=${failureEventId}`,
-    );
-    this.name = "MultipartCompletionFailureReportLeaseLostError";
   }
 }
 
@@ -375,36 +310,6 @@ function requireClaimedJob(
   }
 }
 
-function requireClaimedFailureReport(
-  report: ClaimedMultipartCompletionFailureReport,
-): void {
-  requireUuid(report.failureEventId, "failureEventId");
-  requireUuid(report.attemptToken, "attemptToken");
-  requireWorkspaceId(report.workspaceId);
-  requireUuid(report.leaseToken, "leaseToken");
-  if (
-    report.leaseOwner !== report.leaseOwner.trim()
-    || report.leaseOwner.length < 1
-    || report.leaseOwner.length > 200
-    || controlCharacterPattern.test(report.leaseOwner)
-  ) {
-    throw new TypeError("failure report leaseOwner is invalid.");
-  }
-  if (
-    Number.isSafeInteger(report.retryCount) === false
-    || report.retryCount < 0
-    || Number.isSafeInteger(report.deliveryAttempt) === false
-    || report.deliveryAttempt < 1
-  ) {
-    throw new RangeError(
-      "Failure report retry and delivery counts must be non-negative safe integers.",
-    );
-  }
-  if (safeErrorCodePattern.test(report.errorCode) === false) {
-    throw new TypeError("Failure report errorCode is invalid.");
-  }
-}
-
 function toClaimedJob(row: ClaimedRow): ClaimedMultipartCompletionReconciliation {
   const normalizationVersion = mediaBlobNormalizationVersions.find(
     (candidate) => candidate === row.normalization_version,
@@ -476,27 +381,6 @@ function toClaimedJob(row: ClaimedRow): ClaimedMultipartCompletionReconciliation
   return job;
 }
 
-function toClaimedFailureReport(
-  row: ClaimedFailureReportRow,
-): ClaimedMultipartCompletionFailureReport {
-  const report: ClaimedMultipartCompletionFailureReport = {
-    failureEventId: row.failure_event_id,
-    attemptToken: row.attempt_token,
-    workspaceId: row.workspace_id,
-    retryCount: row.reconciliation_retry_count,
-    errorCode: row.reconciliation_last_error_code,
-    deliveryAttempt: row.failure_report_delivery_count,
-    leaseToken: row.failure_report_lease_token,
-    leaseOwner: row.failure_report_lease_owner,
-    leaseExpiresAt: toIsoString(
-      row.failure_report_lease_expires_at,
-      "failure_report_lease_expires_at",
-    ),
-  };
-  requireClaimedFailureReport(report);
-  return report;
-}
-
 function validateClaimInput(
   input: Readonly<{
     leaseOwner: string;
@@ -545,23 +429,6 @@ export async function claimMultipartCompletionReconciliations(
     [input.leaseOwner, input.leaseDurationMs, input.limit],
   );
   return result.rows.map(toClaimedJob);
-}
-
-export async function claimMultipartCompletionFailureReports(
-  input: Readonly<{
-    leaseOwner: string;
-    leaseDurationMs: number;
-    limit: number;
-    deadlineAtMs: number;
-  }>,
-): Promise<ReadonlyArray<ClaimedMultipartCompletionFailureReport>> {
-  validateClaimInput(input);
-  const result = await unsafeQueryWithDeadline<ClaimedFailureReportRow>(
-    input.deadlineAtMs,
-    "SELECT * FROM content.claim_media_upload_session_completion_failure_reports($1, $2, $3)",
-    [input.leaseOwner, input.leaseDurationMs, input.limit],
-  );
-  return result.rows.map(toClaimedFailureReport);
 }
 
 export async function renewMultipartCompletionReconciliationLease(
@@ -638,76 +505,6 @@ export async function readMultipartCompletionReconciliationOutcome(
     status: row.reconciliation_status,
     errorCode: row.reconciliation_error_code,
   };
-}
-
-export async function finishMultipartCompletionFailureReport(
-  report: ClaimedMultipartCompletionFailureReport,
-  deadlineAtMs: number,
-): Promise<void> {
-  requireClaimedFailureReport(report);
-  const status = await readStatus(
-    deadlineAtMs,
-    `SELECT content.finish_media_upload_session_completion_failure_report(
-       $1, $2, $3
-     ) AS status`,
-    [report.failureEventId, report.attemptToken, report.leaseToken],
-  );
-  if (status === "reported" || status === "already_reported") return;
-  throw new MultipartCompletionFailureReportLeaseLostError(
-    report.failureEventId,
-  );
-}
-
-function toTerminalFailureDetails(
-  report: ClaimedMultipartCompletionFailureReport,
-): MultipartCompletionReconciliationTerminalFailureDetails {
-  return {
-    failureEventId: report.failureEventId,
-    attemptToken: report.attemptToken,
-    workspaceId: report.workspaceId,
-    retryCount: report.retryCount,
-    errorCode: report.errorCode,
-    deliveryAttempt: report.deliveryAttempt,
-  };
-}
-
-export async function deliverMultipartCompletionFailureReport(
-  report: ClaimedMultipartCompletionFailureReport,
-  deadlineAtMs: number,
-  reportTerminalFailure:
-    MultipartCompletionFailureReportBatchInput["reportTerminalFailure"],
-): Promise<void> {
-  requireClaimedFailureReport(report);
-  await unsafeTransactionWithDeadline(deadlineAtMs, async (executor) => {
-    const lockResult = await executor.query<StatusRow>(
-      `SELECT content.lock_media_upload_session_completion_failure_report_delivery(
-         $1, $2, $3
-       ) AS status`,
-      [report.failureEventId, report.attemptToken, report.leaseToken],
-    );
-    const lockStatus = lockResult.rows[0]?.status;
-    if (lockStatus === "already_reported") return;
-    if (lockStatus !== "ready") {
-      throw new MultipartCompletionFailureReportLeaseLostError(
-        report.failureEventId,
-      );
-    }
-
-    reportTerminalFailure(toTerminalFailureDetails(report));
-    const finishResult = await executor.query<StatusRow>(
-      `SELECT content.finish_media_upload_session_completion_failure_report(
-         $1, $2, $3
-       ) AS status`,
-      [report.failureEventId, report.attemptToken, report.leaseToken],
-    );
-    const finishStatus = finishResult.rows[0]?.status;
-    if (finishStatus === "reported" || finishStatus === "already_reported") {
-      return;
-    }
-    throw new MultipartCompletionFailureReportLeaseLostError(
-      report.failureEventId,
-    );
-  });
 }
 
 export async function rescheduleMultipartCompletionReconciliation(
@@ -1231,171 +1028,5 @@ export async function runMultipartCompletionReconciliationBatch(
   return runMultipartCompletionReconciliationBatchWithDependencies(
     input,
     defaultDependencies,
-  );
-}
-
-export type MultipartCompletionFailureReportProcessorDependencies = Readonly<{
-  claimReportsFn: typeof claimMultipartCompletionFailureReports;
-  deliverReportFn: typeof deliverMultipartCompletionFailureReport;
-  finishReportFn: typeof finishMultipartCompletionFailureReport;
-  nowFn: () => number;
-}>;
-
-function failureReportResult(
-  report: ClaimedMultipartCompletionFailureReport,
-  outcome: MultipartCompletionFailureReportOutcome,
-): MultipartCompletionFailureReportResult {
-  return {
-    failureEventId: report.failureEventId,
-    outcome,
-  };
-}
-
-async function deliverFailureReportWithConfirmation(
-  report: ClaimedMultipartCompletionFailureReport,
-  input: MultipartCompletionFailureReportBatchInput,
-  dependencies: MultipartCompletionFailureReportProcessorDependencies,
-): Promise<MultipartCompletionFailureReportResult> {
-  try {
-    await dependencies.deliverReportFn(
-      report,
-      input.deadlineAtMs,
-      input.reportTerminalFailure,
-    );
-    return failureReportResult(report, "reported");
-  } catch (error) {
-    if (error instanceof MultipartCompletionFailureReportLeaseLostError) {
-      return failureReportResult(report, "lease_lost");
-    }
-    if (error instanceof DatabaseCommitOutcomeUnknownError) {
-      try {
-        await dependencies.finishReportFn(report, input.deadlineAtMs);
-        return failureReportResult(report, "reported");
-      } catch (confirmationError) {
-        if (
-          confirmationError instanceof MultipartCompletionFailureReportLeaseLostError
-        ) {
-          return failureReportResult(report, "lease_lost");
-        }
-        if (
-          input.signal.aborted
-          || confirmationError instanceof DatabaseDeadlineExceededError
-          || confirmationError instanceof DatabaseCommitOutcomeUnknownError
-          || confirmationError instanceof TransientDatabaseHttpError
-          || isTransientDatabaseError(confirmationError)
-        ) {
-          return failureReportResult(report, "ambiguous");
-        }
-        throw confirmationError;
-      }
-    }
-    if (
-      input.signal.aborted
-      || error instanceof DatabaseDeadlineExceededError
-      || error instanceof TransientDatabaseHttpError
-      || isTransientDatabaseError(error)
-    ) {
-      return failureReportResult(report, "ambiguous");
-    }
-    throw error;
-  }
-}
-
-export async function processClaimedMultipartCompletionFailureReportWithDependencies(
-  report: ClaimedMultipartCompletionFailureReport,
-  input: MultipartCompletionFailureReportBatchInput,
-  dependencies: MultipartCompletionFailureReportProcessorDependencies,
-): Promise<MultipartCompletionFailureReportResult> {
-  requireClaimedFailureReport(report);
-  input.signal.throwIfAborted();
-  return deliverFailureReportWithConfirmation(
-    report,
-    input,
-    dependencies,
-  );
-}
-
-function countFailureReportOutcome(
-  results: ReadonlyArray<MultipartCompletionFailureReportResult>,
-  outcome: MultipartCompletionFailureReportOutcome,
-): number {
-  return results.filter((item) => item.outcome === outcome).length;
-}
-
-function toFailureReportBatchResult(
-  results: ReadonlyArray<MultipartCompletionFailureReportResult>,
-): MultipartCompletionFailureReportBatchResult {
-  return {
-    claimed: results.length,
-    ambiguous: countFailureReportOutcome(results, "ambiguous"),
-    leaseLost: countFailureReportOutcome(results, "lease_lost"),
-    reported: countFailureReportOutcome(results, "reported"),
-    results,
-  };
-}
-
-export class MultipartCompletionFailureReportBatchError extends Error {
-  constructor(
-    readonly partialResult: MultipartCompletionFailureReportBatchResult,
-    cause: unknown,
-  ) {
-    super(
-      "Multipart completion failure-report batch failed after processing one or more reports.",
-      { cause },
-    );
-    this.name = "MultipartCompletionFailureReportBatchError";
-  }
-}
-
-export async function runMultipartCompletionFailureReportBatchWithDependencies(
-  input: MultipartCompletionFailureReportBatchInput,
-  dependencies: MultipartCompletionFailureReportProcessorDependencies,
-): Promise<MultipartCompletionFailureReportBatchResult> {
-  const results: Array<MultipartCompletionFailureReportResult> = [];
-  try {
-    while (
-      results.length < input.maximumReports
-      && input.signal.aborted === false
-      && dependencies.nowFn() + minimumNewJobBudgetMs < input.deadlineAtMs
-    ) {
-      const claimed = await dependencies.claimReportsFn({
-        leaseOwner: input.leaseOwner,
-        leaseDurationMs: input.leaseDurationMs,
-        limit: 1,
-        deadlineAtMs: input.deadlineAtMs,
-      });
-      const report = claimed[0];
-      if (report === undefined) break;
-      results.push(
-        await processClaimedMultipartCompletionFailureReportWithDependencies(
-          report,
-          input,
-          dependencies,
-        ),
-      );
-    }
-  } catch (error) {
-    throw new MultipartCompletionFailureReportBatchError(
-      toFailureReportBatchResult(results),
-      error,
-    );
-  }
-  return toFailureReportBatchResult(results);
-}
-
-const defaultFailureReportDependencies:
-MultipartCompletionFailureReportProcessorDependencies = {
-  claimReportsFn: claimMultipartCompletionFailureReports,
-  deliverReportFn: deliverMultipartCompletionFailureReport,
-  finishReportFn: finishMultipartCompletionFailureReport,
-  nowFn: Date.now,
-};
-
-export async function runMultipartCompletionFailureReportBatch(
-  input: MultipartCompletionFailureReportBatchInput,
-): Promise<MultipartCompletionFailureReportBatchResult> {
-  return runMultipartCompletionFailureReportBatchWithDependencies(
-    input,
-    defaultFailureReportDependencies,
   );
 }
